fix no gui build and output
